+++ ElvUI/Modules/DataTexts/BattleStats.lua
@@ -9,8 +9,8 @@ local GetBattlefieldScore = GetBattlefieldScore
 
 local displayString = ''
 local holder = {
-	LEFT = { data = {}, '', _G.HONOR, _G.KILLING_BLOWS },
-	RIGHT = { data = {}, _G.KILLS, _G.DEATHS, '' }
+	LEFT = { data = {}, _G.KILLS, _G.KILLING_BLOWS, _G.DEATHS },
+	RIGHT = { data = {}, _G.DAMAGE, _G.HEALS, _G.HONOR }
 }
 
 DT.BattleStats = holder
@@ -35,10 +35,10 @@ function DT:UPDATE_BATTLEFIELD_SCORE()
 	myIndex = nil
 
 	for i = 1, GetNumBattlefieldScores() do
-		local name, kb, hks, deaths, honor = GetBattlefieldScore(i)
+		local name, kb, hks, deaths, honor, _, _, _, _, _, dmg, heals = GetBattlefieldScore(i)
 		if name == E.myname then
-			LEFT[2], LEFT[3] = E:ShortValue(honor), E:ShortValue(kb)
-			RIGHT[1], RIGHT[2] = E:ShortValue(hks), E:ShortValue(deaths)
+			LEFT[1], LEFT[2], LEFT[3] = E:ShortValue(hks), E:ShortValue(kb), E:ShortValue(deaths)
+			RIGHT[1], RIGHT[2], RIGHT[3] = E:ShortValue(dmg), E:ShortValue(heals), E:ShortValue(honor)
 			myIndex = i
 			break
 		end
@@ -69,7 +69,7 @@ function DT:ToggleBattleStats()
 end
 
 local function ValueColorUpdate(hex)
-	displayString = strjoin("", "%s: ", hex, "%s|r")
+	displayString = strjoin('', '%s: ', hex, '%s|r')
 
 	if DT.ShowingBattleStats then
 		DT:UpdateBattlePanel('LEFT')