+++ ElvUI/Modules/Tooltip/Tooltip.lua
@@ -456,7 +456,7 @@ end
 function TT:CheckBackdropColor(tt)
 	if not tt or tt:IsForbidden() then return end
 
-	local r, g, b = E:GetBackdropColor(tt)
+	local r, g, b = tt:GetBackdropColor()
 	if r and g and b then
 		r, g, b = E:Round(r, 1), E:Round(g, 1), E:Round(b, 1)
 
@@ -493,7 +493,7 @@ function TT:SetStyle(tt)
 	if not tt or (tt == E.ScanTooltip or tt.IsEmbedded) or tt:IsForbidden() then return end
 	tt:SetTemplate('Transparent', nil, true) --ignore updates
 
-	local r, g, b = E:GetBackdropColor(tt)
+	local r, g, b = tt:GetBackdropColor()
 	tt:SetBackdropColor(r, g, b, TT.db.colorAlpha)
 end