+++ ElvUI/Settings/Profile.lua
@@ -73,6 +73,7 @@ P.general = {
 			}
 		}
 	},
+	objectiveTracker = true,
 	totems = {
 		enable = true,
 		growthDirection = 'VERTICAL',
@@ -183,7 +184,6 @@ P.bags = {
 	clearSearchOnClose = false,
 	disableBagSort = false,
 	disableBankSort = false,
-	showAssignedColor = true,
 	strata = 'HIGH',
 	qualityColors = true,
 	specialtyColors = true,